Separate runTrigger from runNodeWithInputs
